Proper connection and parsing threading.
The connection to the (currently) fixated json supplier and the code that updates the display components to a threaded handler. Currently updates every 5 seconds.
Planning on making it so that, like the desktop version, the variables will change color based on if it's less (red) or more (green) by simply checking the variable differences (Just a theory but I could probably just make a custom TextView that's for numbers, upon changing the text it'd automatically diffing.)

TODO:
Currently the thread never exists upon full application exit (Really interesting part about this is it actually removes it from the task manager aswell (on 2.2.2 atleast), have to make it interrupt and end the thread.)
As aforementioned I have to aswell make it check differences in the numbers obtained and color them accordingly (Hopefully a fade affect, more than likely possibly.)
When switching to the "more statistics" section I'm thinking of making it instantly obtain the json data again. It's either this or (not currently sure how) make it so that the json code works as intended and doesn't need that screen active in order to have th - wait, I'll just store the values in the main activity and upon creation also initalize it with that - huzzah!
Still have to restructure the table data... /sigh.