Worker table acts as intended.
The table works by using two tables, one table is just the affixed
header, the next is a table with also the header inside of it, used as a
stub for the table alignment, then programmatically the workers will be
added. The scroll view works as intended aswell, the header will always
be there, the workers are scrollable, and the header data even matches
up with the worker data.
ProgressBar is smaller as before it seemed obnoxiously big.
Layout polishing for the miner statistics is indefinitely finished.